Escape event hook
