+++ packages/ui/src/common/hooks/useEscape.ts
@@ -1,0 +1,14 @@
+import { useEffect } from 'react'
+
+export function useEscape(onEscape: () => void) {
+  useEffect(() => {
+    const escapeEvent = (event: KeyboardEvent) => {
+      if (event.key === 'Escape') {
+        onEscape()
+      }
+    }
+    document.addEventListener('keydown', escapeEvent)
+
+    return () => document.removeEventListener('keydown', escapeEvent)
+  }, [])
+}
